Implement history sync on first login
This commit implements history synchronization/back-fill on first login,
as allowed by WhatsApp. Currently, only history for group-chats/MUCs is
allowed, as 1:1 history synchronization is not operational without core
changes/support/privileges for the user's MAM.

NOTE: This commit is experimental, and may or may not break your
sessions. Do not use on production environments.