media: Pass source MIME type to `Convert` call
Calls to convert media will generally be given the (raw) source data and
the target specification for the intended output; internally, we ensure
that target media can be converted appropriately based on the source
MIME type, which is inferred on-demand.

This commit adds the ability for callers to (optionally) specify the
source MIME type in the media specification, which is used to enable
certain edge-cases; e.g. how some MP4 audio files are incorrectly
inferred as `video/mp4`, while not containing any video tracks.