ci: cronjob to automatically bump python dependencies
This will mostly make sure the lockfile is updated to reflect the latest
slidge core compatible version, which ensures the container tagged :main
runs the latest slidge core version too (build is triggered on push),
without needing to manual trigger anything.

It is already in place for slidgram and matteridge and seems to work
fine so far.