Implement basic, read-only support for polls
Initial poll creation by official WhatsApp clients will now be processed
as a plain-text message on the XMPP side, allowing us to surface
interactions such as replies, reactions, etc.

Support for processing any kind of voting, whether incoming or outgoing,
or for getting poll results, is completely missing; this is due to a
number of reasons, not least because XMPP doesn't support polls in any
concrete manner, but also because our upstream WhatsApp client library
has limited support for poll vote update processing.

In the future, and once issues above are alleviated, poll support should
be updated to allow for interactions with polls, both incoming and
outgoing.

Closes: #6