feat: Skip existing historical messages
Messages pulled from history may or may not be useful for remote XMPP
clients; usually, the message archive stored by the XMPP server is the
source-of-truth for what gets pushed out to clients.

This commit implements a simple method of skipping message sends where
existing messages are present in our XMPP archive.

Closes: #3