imprv: consider contact "online" when they interact via WhatsApp client
Before this commit, we relied solely on receiving WhatsApp presences to
change the puppet presence. However, the WhatsApp presence model has
diverged from XMPP: we have to explicitly subscribe to contact presences
and after an unknown amount of time, they stop coming through. Some
logic is already implemented to work around that limitation, by
resubscribing periodically, but this proved to be not enough in
practice.

With this commit, when we receive an event that was triggered by a
contact interaction with the whatsapp client, we send an "online"
presence from this contact, avoiding a possible discrepancy between the
"last seen" info attached to presences, and the last message sent by
this contact.