fix: handling of new contacts
This commit incorporates the new Actor concept in the Contact struct,
allowing actor metadata to be passed to slidge core, even for anonymous
participants identified by a LID only. It may also fix issues relating
to missing metadata (name and avatar) of contacts that were not known at
startup. This is hard to test because I have a very limited number of
whatsapp account I can use to test such things, and given how opaque
contact management is, server-side.

Fixes: https://codeberg.org/slidge/slidge-whatsapp/issues/78